+++ packages/components-js/tests/e2e/specs/banner.e2e.ts
@@ -163,18 +163,15 @@ describe('close', () => {
 describe('lifecycle', () => {
   it('should work without unnecessary round trips on init', async () => {
     await initBanner('error');
-
     const status = await getLifecycleStatus(page);
 
     expect(status.componentDidLoad['p-banner'], 'componentDidLoad: p-banner').toBe(1);
     expect(status.componentDidLoad['p-inline-notification'], 'componentDidLoad: p-inline-notification').toBe(1);
     expect(status.componentDidLoad['p-content-wrapper'], 'componentDidLoad: p-content-wrapper').toBe(1);
-    expect(status.componentDidLoad['p-headline'], 'componentDidLoad: p-headline').toBe(1);
-    expect(status.componentDidLoad['p-text'], 'componentDidLoad: p-text').toBe(2); // one included in button-pure
     expect(status.componentDidLoad['p-icon'], 'componentDidLoad: p-icon').toBe(2); // one included in button-pure
     expect(status.componentDidLoad['p-button-pure'], 'componentDidLoad: p-button-pure').toBe(1);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(9);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(6);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(0);
   });
 
@@ -191,7 +188,7 @@ describe('lifecycle', () => {
     expect(status.componentDidUpdate['p-inline-notification'], 'componentDidUpdate: p-inline-notification').toBe(1);
     expect(status.componentDidUpdate['p-icon'], 'componentDidUpdate: p-icon').toBe(1);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(9);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(6);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(3);
   });
 });