+++ packages/components-js/tests/e2e/specs/carousel.e2e.ts
@@ -494,12 +494,11 @@ describe('lifecycle', () => {
 
     expect(status.componentDidLoad['p-carousel'], 'componentDidLoad: p-carousel').toBe(1);
     expect(status.componentDidLoad['p-button-pure'], 'componentDidLoad: p-button-pure').toBe(2);
-    expect(status.componentDidLoad['p-text'], 'componentDidLoad: p-text').toBe(2);
     expect(status.componentDidLoad['p-icon'], 'componentDidLoad: p-icon').toBe(2);
 
     expect(status.componentDidUpdate['p-button-pure'], 'componentDidUpdate: p-button-pure').toBe(2);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(2);
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(7);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(5);
   });
 
   it('should work without unnecessary round trips on btn next click', async () => {
@@ -512,7 +511,7 @@ describe('lifecycle', () => {
 
     expect(status.componentDidUpdate['p-button-pure'], 'componentDidUpdate: p-button-pure').toBe(4);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(4);
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(7);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(5);
   });
 });