+++ packages/components-js/tests/e2e/specs/link.e2e.ts
@@ -214,7 +214,6 @@ describe('lifecycle', () => {
     const status = await getLifecycleStatus(page);
 
     expect(status.componentDidUpdate['p-link'], 'componentDidUpdate: p-link').toBe(1);
-
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(1);
   });
 });