+++ packages/components-js/tests/e2e/specs/text.e2e.ts
@@ -18,9 +18,9 @@ const initText = (): Promise<void> => {
   return setContentWithDesignSystem(
     page,
     `
-      <p-text>
-        <p>Some message with a <a onclick="return false;" href="#">link</a>.</p>
-      </p-text>`
+    <p-text>
+      <p>Some message with a <a onclick="return false;" href="#">link</a>.</p>
+    </p-text>`
   );
 };
 
@@ -44,7 +44,6 @@ describe('lifecycle', () => {
 
     await setProperty(host, 'weight', 'semibold');
     await waitForStencilLifecycle(page);
-
     const status = await getLifecycleStatus(page);
 
     expect(status.componentDidUpdate['p-text'], 'componentDidUpdate: p-text').toBe(1);
@@ -77,9 +76,9 @@ it('should have "text-size-adjust: none" set', async () => {
   await setContentWithDesignSystem(
     page,
     `
-      <p-text>
-        Some message
-      </p-text>`
+    <p-text>
+      Some message
+    </p-text>`
   );
   const paragraph = await getParagraph();
   const webkitTextSizeAdjustStyle = await getElementStyle(paragraph, 'webkitTextSizeAdjust' as any);