+++ packages/components/src/utils/slotted-styles.ts
@@ -1,4 +1,4 @@
-import { getTagName } from './tag-name';
+import { getTagName } from '.';
 
 type HTMLElementOrDocument = HTMLElement | Document;
 type ElementMap = Map<HTMLElementOrDocument, boolean>;