+++ packages/components/src/utils/validation/throwIfChildrenAreNotOfKind.spec.ts
@@ -1,10 +1,11 @@
+import type { TagName } from '@porsche-design-system/shared';
 import { throwIfChildrenAreNotOfKind } from './throwIfChildrenAreNotOfKind';
 import * as areAllChildrenOfKindUtils from '../dom/areAllChildrenOfKind';
 
 it('should call areChildrenOfKind() with correct parameters', () => {
   const spy = jest.spyOn(areAllChildrenOfKindUtils, 'areAllChildrenOfKind');
   const parent = document.createElement('p-grid');
-  const allowedTagName = 'pGridItem';
+  const allowedTagName: TagName = 'p-grid-item';
 
   throwIfChildrenAreNotOfKind(parent, allowedTagName);
 
@@ -16,10 +17,9 @@ it('should throw correct error message when areChildrenOfKind() is false', () =>
   const parent = document.createElement('p-grid');
   const child1 = document.createElement('p-grid-item');
   const child2 = document.createElement('div');
-  parent.appendChild(child1);
-  parent.appendChild(child2);
+  parent.append(child1, child2);
 
-  expect(() => throwIfChildrenAreNotOfKind(parent, 'pGridItem')).toThrow(
+  expect(() => throwIfChildrenAreNotOfKind(parent, 'p-grid-item')).toThrow(
     'Child HTMLElements of p-grid should be of kind p-grid-item but got div'
   );
 });
@@ -28,5 +28,5 @@ it('should not throw error when areChildrenOfKind() is true', () => {
   jest.spyOn(areAllChildrenOfKindUtils, 'areAllChildrenOfKind').mockReturnValue(true);
   const parent = document.createElement('p-grid');
 
-  expect(() => throwIfChildrenAreNotOfKind(parent, 'pGridItem')).not.toThrow();
+  expect(() => throwIfChildrenAreNotOfKind(parent, 'p-grid-item')).not.toThrow();
 });